flask-backend/requirements.txt
Flask
flask_pymongo
python-dotenv
Requests
numpy
layernext-beta
pymilvus
scikit-learn
tqdm
matplotlib